:pick: +`defragable`
Intended to be used on arena objects where bringing the arena's
implementation internal to a collection's implementation facilitates
defragmentation of the arena. This works as long as the pointers
associated to the object are internal to the implementation and the user
does not directly access them. For example, `vector` internally handles
its array, and the user does not directly access the array, so a static
arena allocator within `vector` can change the pointers around of
allocated arrays without affecting the user.